requirements.txt
streamlit
nest_asyncio
python-dotenv
langchain-groq
streamlit-mic-recorder
SpeechRecognition
mcp[cli]
mcp-use
requests
langgraph
langchain
